Move escape method to utils.js
